+++ modules/ravestate_ui_new/src/app/app.module.ts
@@ -2,9 +2,9 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 
 import { AppComponent } from './app.component';
-import { NodeComponent } from './svg-experiments/elements/node.component';
-import { SvgExperimentsComponent } from './svg-experiments/svg-experiments.component';
-import { ConnectorComponent } from './svg-experiments/elements/connector.component';
+import { NodeComponent } from './svg/elements/node.component';
+import { StaticGraphExampleComponent } from './svg/static-graph-example/static-graph-example.component';
+import { ConnectorComponent } from './svg/elements/connector.component';
 import { MockDataControllerComponent } from './mock-data/mock-data-controller/mock-data-controller.component';
 
 @NgModule({
@@ -12,7 +12,7 @@ import { MockDataControllerComponent } from './mock-data/mock-data-controller/mo
         AppComponent,
         NodeComponent,
         ConnectorComponent,
-        SvgExperimentsComponent,
+        StaticGraphExampleComponent,
         MockDataControllerComponent
     ],
     imports: [
