+++ modules/ravestate_ui_new/src/app/svg/elements/connector.component.ts
