+++ modules/ravestate_ui_new/src/app/svg/elements/node.component.ts
