+++ modules/ravestate_ui_new/src/app/svg/elements/styles.scss
