+++ modules/ravestate_ui_new/src/app/svg/static-graph-example/static-graph-example.component.scss
